Fixed all remaining compiler warnings

+++ Trogdor-Reburninated/src/sprite_objects.h
@@ -20,18 +20,18 @@ extern void DRAW_RECT_WITH_ALPHA(SDL_Rect, Uint8, Uint8, Uint8, Uint8);
 
 #define PREPARE_SPRITE(spriteObj, path, rect_x, rect_y, numSprites_x, numSprites_y, scale) \
     prepareSprite(&spriteObj, path, numSprites_x, numSprites_y, scale);                    \
-    setSpritePos(&spriteObj, rect_x, rect_y);
+    setSpritePos(&spriteObj, (int)rect_x, (int)rect_y);
 
 #define OBJ_TO_MID_SCREEN_X(obj) \
     ((gameWidth - obj.dstrect.w) / 2)
 
 #define OBJ_TO_SCREEN_AT_FRACTION_X(obj, val) \
-    ((gameWidth * val) - (obj.dstrect.w / 2))
+    (((double)gameWidth * val) - (obj.dstrect.w / 2))
 
 #define OBJ_TO_MID_SCREEN_Y(obj) \
     ((gameHeight - obj.dstrect.h) / 2)
 
 #define OBJ_TO_SCREEN_AT_FRACTION_Y(obj, val) \
-    ((gameHeight * val) - (obj.dstrect.h / 2))
+    (((double)gameHeight * val) - (obj.dstrect.h / 2))
 
 #endif
